feat: Dockerfile to run Juputer server in a container
This Dockerfile us used in mlcourse.ai and closely follows versions
of typical data processing libraries that are used in Kaggle Docker image.

The docker file itself has no entry point and thus can not be used in
a simple docker run command. It is used with scripts found in the same dir